+++ Classes/Event/TrackSiteSearchEvent.php
@@ -12,9 +12,11 @@ declare(strict_types=1);
 namespace Brotkrueml\MatomoIntegration\Event;
 
 use Brotkrueml\MatomoIntegration\Entity\CustomDimension;
+use Psr\Http\Message\ServerRequestInterface;
 
 final class TrackSiteSearchEvent
 {
+    private ServerRequestInterface $request;
     private string $keyword = '';
     /**
      * @var string|false
@@ -28,6 +30,16 @@ final class TrackSiteSearchEvent
      * @var CustomDimension[]
      */
     private array $customDimensions = [];
+
+    public function __construct(ServerRequestInterface $request)
+    {
+        $this->request = $request;
+    }
+
+    public function getRequest(): ServerRequestInterface
+    {
+        return $this->request;
+    }
 
     /**
      * @internal